+++ PULLCON0.9/PULLCON0.9/source/maingame/SupplyEria.h
@@ -17,6 +17,6 @@ public:
 
 protected:
 	int    _handle;
-	int radius;
+	float radius;
 };
